+++ pkg/cluster/deploystorage_resources.go
@@ -4,10 +4,15 @@ package cluster
 // Licensed under the Apache License 2.0.
 
 import (
+	"fmt"
+
+	mgmtnetwork "github.com/Azure/azure-sdk-for-go/services/network/mgmt/2019-07-01/network"
 	mgmtauthorization "github.com/Azure/azure-sdk-for-go/services/preview/authorization/mgmt/2018-09-01-preview/authorization"
 	mgmtstorage "github.com/Azure/azure-sdk-for-go/services/storage/mgmt/2019-04-01/storage"
 	"github.com/Azure/go-autorest/autorest/to"
+	"github.com/openshift/installer/pkg/asset/installconfig"
 
+	"github.com/Azure/ARO-RP/pkg/api"
 	"github.com/Azure/ARO-RP/pkg/util/arm"
 	"github.com/Azure/ARO-RP/pkg/util/azureclient"
 	"github.com/Azure/ARO-RP/pkg/util/feature"
@@ -107,6 +112,326 @@ func (m *manager) clusterStorageAccountBlob(name string) *arm.Resource {
 		APIVersion: azureclient.APIVersion("Microsoft.Storage"),
 		DependsOn: []string{
 			"Microsoft.Storage/storageAccounts/cluster" + m.doc.OpenShiftCluster.Properties.StorageSuffix,
+		},
+	}
+}
+
+func (m *manager) networkPrivateLinkService(installConfig *installconfig.InstallConfig) *arm.Resource {
+	return &arm.Resource{
+		Resource: &mgmtnetwork.PrivateLinkService{
+			PrivateLinkServiceProperties: &mgmtnetwork.PrivateLinkServiceProperties{
+				LoadBalancerFrontendIPConfigurations: &[]mgmtnetwork.FrontendIPConfiguration{
+					{
+						ID: to.StringPtr(fmt.Sprintf("[resourceId('Microsoft.Network/loadBalancers/frontendIPConfigurations', '%s-internal', 'internal-lb-ip-v4')]", m.doc.OpenShiftCluster.Properties.InfraID)),
+					},
+				},
+				IPConfigurations: &[]mgmtnetwork.PrivateLinkServiceIPConfiguration{
+					{
+						PrivateLinkServiceIPConfigurationProperties: &mgmtnetwork.PrivateLinkServiceIPConfigurationProperties{
+							Subnet: &mgmtnetwork.Subnet{
+								ID: to.StringPtr(m.doc.OpenShiftCluster.Properties.MasterProfile.SubnetID),
+							},
+						},
+						Name: to.StringPtr(m.doc.OpenShiftCluster.Properties.InfraID + "-pls-nic"),
+					},
+				},
+				Visibility: &mgmtnetwork.PrivateLinkServicePropertiesVisibility{
+					Subscriptions: &[]string{
+						m.env.SubscriptionID(),
+					},
+				},
+				AutoApproval: &mgmtnetwork.PrivateLinkServicePropertiesAutoApproval{
+					Subscriptions: &[]string{
+						m.env.SubscriptionID(),
+					},
+				},
+			},
+			Name:     to.StringPtr(m.doc.OpenShiftCluster.Properties.InfraID + "-pls"),
+			Type:     to.StringPtr("Microsoft.Network/privateLinkServices"),
+			Location: &installConfig.Config.Azure.Region,
+		},
+		APIVersion: azureclient.APIVersion("Microsoft.Network"),
+		DependsOn: []string{
+			"Microsoft.Network/loadBalancers/" + m.doc.OpenShiftCluster.Properties.InfraID + "-internal",
+		},
+	}
+}
+
+func (m *manager) networkPublicIPAddress(installConfig *installconfig.InstallConfig, name string) *arm.Resource {
+	return &arm.Resource{
+		Resource: &mgmtnetwork.PublicIPAddress{
+			Sku: &mgmtnetwork.PublicIPAddressSku{
+				Name: mgmtnetwork.PublicIPAddressSkuNameStandard,
+			},
+			PublicIPAddressPropertiesFormat: &mgmtnetwork.PublicIPAddressPropertiesFormat{
+				PublicIPAllocationMethod: mgmtnetwork.Static,
+			},
+			Name:     &name,
+			Type:     to.StringPtr("Microsoft.Network/publicIPAddresses"),
+			Location: &installConfig.Config.Azure.Region,
+		},
+		APIVersion: azureclient.APIVersion("Microsoft.Network"),
+	}
+}
+
+func (m *manager) networkInternalLoadBalancer(installConfig *installconfig.InstallConfig) *arm.Resource {
+	return &arm.Resource{
+		Resource: &mgmtnetwork.LoadBalancer{
+			Sku: &mgmtnetwork.LoadBalancerSku{
+				Name: mgmtnetwork.LoadBalancerSkuNameStandard,
+			},
+			LoadBalancerPropertiesFormat: &mgmtnetwork.LoadBalancerPropertiesFormat{
+				FrontendIPConfigurations: &[]mgmtnetwork.FrontendIPConfiguration{
+					{
+						FrontendIPConfigurationPropertiesFormat: &mgmtnetwork.FrontendIPConfigurationPropertiesFormat{
+							PrivateIPAllocationMethod: mgmtnetwork.Dynamic,
+							Subnet: &mgmtnetwork.Subnet{
+								ID: to.StringPtr(m.doc.OpenShiftCluster.Properties.MasterProfile.SubnetID),
+							},
+						},
+						Name: to.StringPtr("internal-lb-ip-v4"),
+					},
+				},
+				BackendAddressPools: &[]mgmtnetwork.BackendAddressPool{
+					{
+						Name: to.StringPtr(m.doc.OpenShiftCluster.Properties.InfraID),
+					},
+					{
+						Name: to.StringPtr("ssh-0"),
+					},
+					{
+						Name: to.StringPtr("ssh-1"),
+					},
+					{
+						Name: to.StringPtr("ssh-2"),
+					},
+				},
+				LoadBalancingRules: &[]mgmtnetwork.LoadBalancingRule{
+					{
+						LoadBalancingRulePropertiesFormat: &mgmtnetwork.LoadBalancingRulePropertiesFormat{
+							FrontendIPConfiguration: &mgmtnetwork.SubResource{
+								ID: to.StringPtr(fmt.Sprintf("[resourceId('Microsoft.Network/loadBalancers/frontendIPConfigurations', '%s-internal', 'internal-lb-ip-v4')]", m.doc.OpenShiftCluster.Properties.InfraID)),
+							},
+							BackendAddressPool: &mgmtnetwork.SubResource{
+								ID: to.StringPtr(fmt.Sprintf("[resourceId('Microsoft.Network/loadBalancers/backendAddressPools', '%s-internal', '%[1]s')]", m.doc.OpenShiftCluster.Properties.InfraID)),
+							},
+							Probe: &mgmtnetwork.SubResource{
+								ID: to.StringPtr(fmt.Sprintf("[resourceId('Microsoft.Network/loadBalancers/probes', '%s-internal', 'api-internal-probe')]", m.doc.OpenShiftCluster.Properties.InfraID)),
+							},
+							Protocol:             mgmtnetwork.TransportProtocolTCP,
+							LoadDistribution:     mgmtnetwork.LoadDistributionDefault,
+							FrontendPort:         to.Int32Ptr(6443),
+							BackendPort:          to.Int32Ptr(6443),
+							IdleTimeoutInMinutes: to.Int32Ptr(30),
+							DisableOutboundSnat:  to.BoolPtr(true),
+						},
+						Name: to.StringPtr("api-internal-v4"),
+					},
+					{
+						LoadBalancingRulePropertiesFormat: &mgmtnetwork.LoadBalancingRulePropertiesFormat{
+							FrontendIPConfiguration: &mgmtnetwork.SubResource{
+								ID: to.StringPtr(fmt.Sprintf("[resourceId('Microsoft.Network/loadBalancers/frontendIPConfigurations', '%s-internal', 'internal-lb-ip-v4')]", m.doc.OpenShiftCluster.Properties.InfraID)),
+							},
+							BackendAddressPool: &mgmtnetwork.SubResource{
+								ID: to.StringPtr(fmt.Sprintf("[resourceId('Microsoft.Network/loadBalancers/backendAddressPools', '%s-internal', '%[1]s')]", m.doc.OpenShiftCluster.Properties.InfraID)),
+							},
+							Probe: &mgmtnetwork.SubResource{
+								ID: to.StringPtr(fmt.Sprintf("[resourceId('Microsoft.Network/loadBalancers/probes', '%s-internal', 'sint-probe')]", m.doc.OpenShiftCluster.Properties.InfraID)),
+							},
+							Protocol:             mgmtnetwork.TransportProtocolTCP,
+							LoadDistribution:     mgmtnetwork.LoadDistributionDefault,
+							FrontendPort:         to.Int32Ptr(22623),
+							BackendPort:          to.Int32Ptr(22623),
+							IdleTimeoutInMinutes: to.Int32Ptr(30),
+						},
+						Name: to.StringPtr("sint-v4"),
+					},
+					{
+						LoadBalancingRulePropertiesFormat: &mgmtnetwork.LoadBalancingRulePropertiesFormat{
+							FrontendIPConfiguration: &mgmtnetwork.SubResource{
+								ID: to.StringPtr(fmt.Sprintf("[resourceId('Microsoft.Network/loadBalancers/frontendIPConfigurations', '%s-internal', 'internal-lb-ip-v4')]", m.doc.OpenShiftCluster.Properties.InfraID)),
+							},
+							BackendAddressPool: &mgmtnetwork.SubResource{
+								ID: to.StringPtr(fmt.Sprintf("[resourceId('Microsoft.Network/loadBalancers/backendAddressPools', '%s-internal', 'ssh-0')]", m.doc.OpenShiftCluster.Properties.InfraID)),
+							},
+							Probe: &mgmtnetwork.SubResource{
+								ID: to.StringPtr(fmt.Sprintf("[resourceId('Microsoft.Network/loadBalancers/probes', '%s-internal', 'ssh')]", m.doc.OpenShiftCluster.Properties.InfraID)),
+							},
+							Protocol:             mgmtnetwork.TransportProtocolTCP,
+							LoadDistribution:     mgmtnetwork.LoadDistributionDefault,
+							FrontendPort:         to.Int32Ptr(2200),
+							BackendPort:          to.Int32Ptr(22),
+							IdleTimeoutInMinutes: to.Int32Ptr(30),
+							DisableOutboundSnat:  to.BoolPtr(true),
+						},
+						Name: to.StringPtr("ssh-0"),
+					},
+					{
+						LoadBalancingRulePropertiesFormat: &mgmtnetwork.LoadBalancingRulePropertiesFormat{
+							FrontendIPConfiguration: &mgmtnetwork.SubResource{
+								ID: to.StringPtr(fmt.Sprintf("[resourceId('Microsoft.Network/loadBalancers/frontendIPConfigurations', '%s-internal', 'internal-lb-ip-v4')]", m.doc.OpenShiftCluster.Properties.InfraID)),
+							},
+							BackendAddressPool: &mgmtnetwork.SubResource{
+								ID: to.StringPtr(fmt.Sprintf("[resourceId('Microsoft.Network/loadBalancers/backendAddressPools', '%s-internal', 'ssh-1')]", m.doc.OpenShiftCluster.Properties.InfraID)),
+							},
+							Probe: &mgmtnetwork.SubResource{
+								ID: to.StringPtr(fmt.Sprintf("[resourceId('Microsoft.Network/loadBalancers/probes', '%s-internal', 'ssh')]", m.doc.OpenShiftCluster.Properties.InfraID)),
+							},
+							Protocol:             mgmtnetwork.TransportProtocolTCP,
+							LoadDistribution:     mgmtnetwork.LoadDistributionDefault,
+							FrontendPort:         to.Int32Ptr(2201),
+							BackendPort:          to.Int32Ptr(22),
+							IdleTimeoutInMinutes: to.Int32Ptr(30),
+							DisableOutboundSnat:  to.BoolPtr(true),
+						},
+						Name: to.StringPtr("ssh-1"),
+					},
+					{
+						LoadBalancingRulePropertiesFormat: &mgmtnetwork.LoadBalancingRulePropertiesFormat{
+							FrontendIPConfiguration: &mgmtnetwork.SubResource{
+								ID: to.StringPtr(fmt.Sprintf("[resourceId('Microsoft.Network/loadBalancers/frontendIPConfigurations', '%s-internal', 'internal-lb-ip-v4')]", m.doc.OpenShiftCluster.Properties.InfraID)),
+							},
+							BackendAddressPool: &mgmtnetwork.SubResource{
+								ID: to.StringPtr(fmt.Sprintf("[resourceId('Microsoft.Network/loadBalancers/backendAddressPools', '%s-internal', 'ssh-2')]", m.doc.OpenShiftCluster.Properties.InfraID)),
+							},
+							Probe: &mgmtnetwork.SubResource{
+								ID: to.StringPtr(fmt.Sprintf("[resourceId('Microsoft.Network/loadBalancers/probes', '%s-internal', 'ssh')]", m.doc.OpenShiftCluster.Properties.InfraID)),
+							},
+							Protocol:             mgmtnetwork.TransportProtocolTCP,
+							LoadDistribution:     mgmtnetwork.LoadDistributionDefault,
+							FrontendPort:         to.Int32Ptr(2202),
+							BackendPort:          to.Int32Ptr(22),
+							IdleTimeoutInMinutes: to.Int32Ptr(30),
+							DisableOutboundSnat:  to.BoolPtr(true),
+						},
+						Name: to.StringPtr("ssh-2"),
+					},
+				},
+				Probes: &[]mgmtnetwork.Probe{
+					{
+						ProbePropertiesFormat: &mgmtnetwork.ProbePropertiesFormat{
+							Protocol:          mgmtnetwork.ProbeProtocolHTTPS,
+							Port:              to.Int32Ptr(6443),
+							IntervalInSeconds: to.Int32Ptr(5),
+							NumberOfProbes:    to.Int32Ptr(2),
+							RequestPath:       to.StringPtr("/readyz"),
+						},
+						Name: to.StringPtr("api-internal-probe"),
+					},
+					{
+						ProbePropertiesFormat: &mgmtnetwork.ProbePropertiesFormat{
+							Protocol:          mgmtnetwork.ProbeProtocolHTTPS,
+							Port:              to.Int32Ptr(22623),
+							IntervalInSeconds: to.Int32Ptr(5),
+							NumberOfProbes:    to.Int32Ptr(2),
+							RequestPath:       to.StringPtr("/healthz"),
+						},
+						Name: to.StringPtr("sint-probe"),
+					},
+					{
+						ProbePropertiesFormat: &mgmtnetwork.ProbePropertiesFormat{
+							Protocol:          mgmtnetwork.ProbeProtocolTCP,
+							Port:              to.Int32Ptr(22),
+							IntervalInSeconds: to.Int32Ptr(5),
+							NumberOfProbes:    to.Int32Ptr(2),
+						},
+						Name: to.StringPtr("ssh"),
+					},
+				},
+			},
+			Name:     to.StringPtr(m.doc.OpenShiftCluster.Properties.InfraID + "-internal"),
+			Type:     to.StringPtr("Microsoft.Network/loadBalancers"),
+			Location: &installConfig.Config.Azure.Region,
+		},
+		APIVersion: azureclient.APIVersion("Microsoft.Network"),
+	}
+}
+
+func (m *manager) networkPublicLoadBalancer(installConfig *installconfig.InstallConfig) *arm.Resource {
+	lb := &mgmtnetwork.LoadBalancer{
+		Sku: &mgmtnetwork.LoadBalancerSku{
+			Name: mgmtnetwork.LoadBalancerSkuNameStandard,
+		},
+		LoadBalancerPropertiesFormat: &mgmtnetwork.LoadBalancerPropertiesFormat{
+			FrontendIPConfigurations: &[]mgmtnetwork.FrontendIPConfiguration{
+				{
+					FrontendIPConfigurationPropertiesFormat: &mgmtnetwork.FrontendIPConfigurationPropertiesFormat{
+						PublicIPAddress: &mgmtnetwork.PublicIPAddress{
+							ID: to.StringPtr("[resourceId('Microsoft.Network/publicIPAddresses', '" + m.doc.OpenShiftCluster.Properties.InfraID + "-pip-v4')]"),
+						},
+					},
+					Name: to.StringPtr("public-lb-ip-v4"),
+				},
+			},
+			BackendAddressPools: &[]mgmtnetwork.BackendAddressPool{
+				{
+					Name: to.StringPtr(m.doc.OpenShiftCluster.Properties.InfraID),
+				},
+			},
+			LoadBalancingRules: &[]mgmtnetwork.LoadBalancingRule{}, //required to override default LB rules for port 80 and 443
+			Probes:             &[]mgmtnetwork.Probe{},             //required to override default LB rules for port 80 and 443
+			OutboundRules: &[]mgmtnetwork.OutboundRule{
+				{
+					OutboundRulePropertiesFormat: &mgmtnetwork.OutboundRulePropertiesFormat{
+						FrontendIPConfigurations: &[]mgmtnetwork.SubResource{
+							{
+								ID: to.StringPtr("[resourceId('Microsoft.Network/loadBalancers/frontendIPConfigurations', '" + m.doc.OpenShiftCluster.Properties.InfraID + "', 'public-lb-ip-v4')]"),
+							},
+						},
+						BackendAddressPool: &mgmtnetwork.SubResource{
+							ID: to.StringPtr(fmt.Sprintf("[resourceId('Microsoft.Network/loadBalancers/backendAddressPools', '%s', '%[1]s')]", m.doc.OpenShiftCluster.Properties.InfraID)),
+						},
+						Protocol:             mgmtnetwork.LoadBalancerOutboundRuleProtocolAll,
+						IdleTimeoutInMinutes: to.Int32Ptr(30),
+					},
+					Name: to.StringPtr("outbound-rule-v4"),
+				},
+			},
+		},
+		Name:     to.StringPtr(m.doc.OpenShiftCluster.Properties.InfraID),
+		Type:     to.StringPtr("Microsoft.Network/loadBalancers"),
+		Location: &installConfig.Config.Azure.Region,
+	}
+
+	if m.doc.OpenShiftCluster.Properties.APIServerProfile.Visibility == api.VisibilityPublic {
+		*lb.LoadBalancingRules = append(*lb.LoadBalancingRules, mgmtnetwork.LoadBalancingRule{
+			LoadBalancingRulePropertiesFormat: &mgmtnetwork.LoadBalancingRulePropertiesFormat{
+				FrontendIPConfiguration: &mgmtnetwork.SubResource{
+					ID: to.StringPtr(fmt.Sprintf("[resourceId('Microsoft.Network/loadBalancers/frontendIPConfigurations', '%s', 'public-lb-ip-v4')]", m.doc.OpenShiftCluster.Properties.InfraID)),
+				},
+				BackendAddressPool: &mgmtnetwork.SubResource{
+					ID: to.StringPtr(fmt.Sprintf("[resourceId('Microsoft.Network/loadBalancers/backendAddressPools', '%s', '%[1]s')]", m.doc.OpenShiftCluster.Properties.InfraID)),
+				},
+				Probe: &mgmtnetwork.SubResource{
+					ID: to.StringPtr(fmt.Sprintf("[resourceId('Microsoft.Network/loadBalancers/probes', '%s', 'api-internal-probe')]", m.doc.OpenShiftCluster.Properties.InfraID)),
+				},
+				Protocol:             mgmtnetwork.TransportProtocolTCP,
+				LoadDistribution:     mgmtnetwork.LoadDistributionDefault,
+				FrontendPort:         to.Int32Ptr(6443),
+				BackendPort:          to.Int32Ptr(6443),
+				IdleTimeoutInMinutes: to.Int32Ptr(30),
+				DisableOutboundSnat:  to.BoolPtr(true),
+			},
+			Name: to.StringPtr("api-internal-v4"),
+		})
+
+		*lb.Probes = append(*lb.Probes, mgmtnetwork.Probe{
+			ProbePropertiesFormat: &mgmtnetwork.ProbePropertiesFormat{
+				Protocol:          mgmtnetwork.ProbeProtocolHTTPS,
+				Port:              to.Int32Ptr(6443),
+				IntervalInSeconds: to.Int32Ptr(5),
+				NumberOfProbes:    to.Int32Ptr(2),
+				RequestPath:       to.StringPtr("/readyz"),
+			},
+			Name: to.StringPtr("api-internal-probe"),
+		})
+	}
+
+	return &arm.Resource{
+		Resource:   lb,
+		APIVersion: azureclient.APIVersion("Microsoft.Network"),
+		DependsOn: []string{
+			"Microsoft.Network/publicIPAddresses/" + m.doc.OpenShiftCluster.Properties.InfraID + "-pip-v4",
 		},
 	}
 }
